requirements.txt
fastapi
uvicorn
groq
python-dotenv
requests
jinja2
